moved loadFeaturesFromPackages to its own file
