+++ packages/scripts/src/analyze-feature/index.ts
@@ -1,1 +1,2 @@
-export { loadFeaturesFromPackages } from './analyze-feature'
+export { loadFeaturesFromPaths } from './analyze-feature'
+export { loadFeaturesFromPackages } from './features-from-packages'
